Add files via upload
So the only things left are the cardson the homepage, they dont look great and to make the nav bars uniform of which you will chose which one is better between the one on the homepage or the one on the dashboard.
additionally i didnt change any color scheme because it appears to be that way and also it suits. Please inform me of what to change so i target them directly. Arigato